Initial work on adding Jetty
