+++ README.md
@@ -20,7 +20,8 @@ Tomcat connector to this port, so the only thing needed on your side is to open 
 
 # Servlet containers supported (embedded)
 
- - Tomcat
+ - [Tomcat](tomcat)
+ - [Jetty](jetty)
  
 # Usage
 
@@ -90,5 +91,5 @@ to your configiration
 
 ## Testing locally
 
-The library has integration tests in [src/test/java](src/test/java) directory. One can adapt these tests
+The library has integration tests in [src/test/java](tomcat/src/test/java) directory. One can adapt these tests
 according to own needs, as they use Pebble - LetsEncrypt testing server.
