+++ build.gradle
@@ -3,48 +3,26 @@ plugins {
     id 'maven-publish'
 }
 
-group = 'com.github.valb3r'
-version = '0.1.5-SNAPSHOT'
-sourceCompatibility = '11'
-
 repositories {
     mavenCentral()
 }
 
-dependencies {
-    compileOnly 'org.springframework:spring-context:5.3.12'
-    compileOnly 'org.springframework:spring-webmvc:5.3.12'
-    compileOnly 'org.springframework.boot:spring-boot-autoconfigure:2.5.6'
-    compileOnly 'org.apache.tomcat.embed:tomcat-embed-core:9.0.38'
+ext {
+    spring = '5.3.13'
+    springBoot = '2.5.6'
+    acme4j = '2.12'
+    tomcatEmbed = '9.0.38'
+    jettyEmbed = '9.4.3.v20170317'
+    bouncycastle = '1.69'
 
-    implementation 'org.shredzone.acme4j:acme4j-client:2.12'
-    implementation 'org.shredzone.acme4j:acme4j-utils:2.12'
-
-    implementation 'org.bouncycastle:bcprov-jdk15on:1.69'
-
-
-    testImplementation 'org.springframework.boot:spring-boot-starter-web:2.5.6'
-    testImplementation 'org.springframework.boot:spring-boot-starter-tomcat:2.5.6'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test:2.5.6'
-
-    testImplementation 'org.awaitility:awaitility:4.1.1'
-    testImplementation 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
-    testImplementation 'org.testcontainers:junit-jupiter:1.16.2'
-    testImplementation 'org.assertj:assertj-core:3.21.0'
+    // tests
+    testContainers = '1.16.2'
+    awaitability = '4.1.1'
+    jupiter = '5.8.1'
+    assertj = '3.21.0'
 }
 
-test {
-    useJUnitPlatform()
-}
 
-publishing {
-    publications {
-        mavenJava(MavenPublication) {
-            groupId = 'com.github.valb3r'
-            artifactId = 'letsencrypt-helper'
-            version = '0.1.5-SNAPSHOT'
-
-            from components.java
-        }
-    }
-}
+group = 'com.github.valb3r'
+version = '0.1.5-SNAPSHOT'
+sourceCompatibility = '11'
