--- src/test/java/com/github/valb3r/letsencrypt/dummyapp/DummyRest.java
@@ -1,12 +1,0 @@
-package com.github.valb3r.letsencrypt.dummyapp;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class DummyRest {
-    @GetMapping("/hello")
-    public String hello() {
-        return "hello";
-    }
-}
