+++ tests-common/src/main/java/com/github/valb3r/letsencrypthelper/dummyapp/DummyRest.java
@@ -1,0 +1,12 @@
+package com.github.valb3r.letsencrypthelper.dummyapp;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class DummyRest {
+    @GetMapping("/hello")
+    public String hello() {
+        return "hello";
+    }
+}
