+++ tests-common/src/main/resources/application-test-missing-keystore.yml
