+++ tests-common/src/main/resources/application-test-not-expired-keystore.yml
@@ -1,0 +1,15 @@
+server:
+  port: 8443
+  ssl:
+    key-store: not-expired-test-keystore
+    key-store-password: change-me # Password for KeyStore protection
+    key-store-type: PKCS12
+    key-alias: tomcat # Certificate name in KeyStore
+    enabled: true # Important to place this explicitly
+lets-encrypt-helper:
+  domain: host.testcontainers.internal # Domain to issue certificate for
+  contact: mailto:john.doe@mymail.example.com # Your contact for LetsEncrypt
+  busy-wait-interval: PT5S
+  letsencrypt-server: acme://pebble
+  development-only:
+    http01-challenge-port: 5002 # Pebble by default will try to challenge this port
