+++ tests-common/src/main/resources/not-expired-test-keystore
