+++ tomcat/src/test/java/com/github/valb3r/letsencrypthelper/TomcatExpiredKeyPebbleTest.java
@@ -1,0 +1,4 @@
+package com.github.valb3r.letsencrypthelper;
+
+class TomcatExpiredKeyPebbleTest extends ExpiredKeyPebbleTest {
+}
