+++ tomcat/src/test/java/com/github/valb3r/letsencrypthelper/TomcatNotExpiredKeystorePebbleTest.java
@@ -1,0 +1,4 @@
+package com.github.valb3r.letsencrypthelper;
+
+class TomcatNotExpiredKeystorePebbleTest extends NotExpiredKeystorePebbleTest {
+}
